Fix type hinting as possible
Type hinting is only being used for documentation at this time,
but it's good to use specific and syntactically correct type hints.
However, to use the type checker in a useful way, you'd have to
use isinstance instead of string tags for the AstNodes
(or an equivalent "is_type(*AstNode)" function).